Move ui notification to custom hook
